페이지 이동시 토큰이 만료되었다면 자동 갱신하는 기능 추가

+++ src/pages/Logout.js
@@ -9,19 +9,26 @@ import query from "api";
 function Page({ ...props }) {
 
     useEffect(()=>{
-        query({
-            url: "/auth/signout",
-            method: "DELETE",
-        })
-        .then((res) => {
-            console.log(res);
-        })
-        .finally((e)=>{
-            console.log("finally", e);
-            store.clearAll();
-            console.log("logout component: logout success");
+        let user = store.get('user');
+        if(!!user && !!user.token && !!user.key) {
+            // 토큰이 있을때만 토큰 만료 처리
+            query({
+                url: "/auth/signout",
+                method: "DELETE",
+            })
+            .then((res) => {
+                console.log(res);
+            })
+            .finally((e)=>{
+                console.log("finally", e);
+                store.clearAll();
+                console.log("logout component: logout success");
+                props.history.push('/login');
+            })
+        } else {
             props.history.push('/login');
-        })
+        }
+        
     }, [])
     
     return null;
